Lots of fixes #10
