+++ TM.Digital.Services/AllActionFinishedEventHandler.cs
@@ -1,0 +1,6 @@
+﻿using TM.Digital.Model.Player;
+
+namespace TM.Digital.Services
+{
+    public delegate void AllActionFinishedEventHandler(Player player);
+}
