+++ TM.Digital.Services/IPrerequisiteStrategy.cs
@@ -1,0 +1,12 @@
+﻿using TM.Digital.Model.Board;
+using TM.Digital.Model.Cards;
+using TM.Digital.Model.Player;
+
+namespace TM.Digital.Services
+{
+    public interface IPrerequisiteStrategy
+
+    {
+        bool CanPlayCard(Patent inputPatent, Board currentBoardState, Player patentOwner);
+    }
+}
